+++ views/pages/index.php
@@ -47,10 +47,24 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
 
                     if (($pageURL = $model->getPageUrl(true, true)) && $model->id)
                         $output .= '<br/>' . Html::a($model->getUrl(true), $pageURL, [
-                                'target' => '_blank',
-                                'data-pjax' => 0
-                            ]);;
+                            'target' => '_blank',
+                            'data-pjax' => 0
+                        ]);
 
+                    if (isset(Yii::$app->redirects) && $model->url && ($model->status == $model::STATUS_PUBLISHED)) {
+                        if ($url = Yii::$app->redirects->check($model->url, false)) {
+                            $output .= '&nbsp' . Html::tag('span', '', [
+                                'class' => "text-danger fa fa-exclamation-circle",
+                                'data' => [
+                                    'toggle' => "tooltip",
+                                    'placement' => "top"
+                                ],
+                                'title' => Yii::t('app/modules/redirects', 'For this URL is active redirect to {url}', [
+                                    'url' => $url
+                                ])
+                            ]);
+                        }
+                    }
                     return $output;
                 }
             ],